requirements.txt
fastapi
uvicorn
transformers>=4.30.0
torch>=2.0.0
Pillow>=10.0.0
reportlab>=4.0.0
python-multipart
accelerate
